Added Contribution Spreadsheet
From Assessment on Git -

Groups must maintain a spreadsheet detailing individual team members contribution at each of the 5 assessment points
We wish to determine the individual contribution to the team project. To do this, the team have to submit a single spreadsheet to Moodle defining the agreed contribution of each team member to the individual delivery points. This should be submitted in percentages with the total sum of individual contributions adding up to 100% at each point of assessment. 

Code review dates may not be accuarate as I am pulling from the Git and some parts are out of date